# AttendanceTrackerInfrastructure/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AttendanceTrackerInfrastructure, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/AttendanceTrackerInfrastructure/get-admins": {
      "get": {
        "tags": [
          "AttendanceTrackerInfrastructure"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AttendanceTrackerInfrastructure/get-admin/{adminName}": {
      "get": {
        "tags": [
          "AttendanceTrackerInfrastructure"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "adminName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AttendanceTrackerInfrastructure/get-staffs": {
      "get": {
        "tags": [
          "AttendanceTrackerInfrastructure"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AttendanceTrackerInfrastructure/get-departments": {
      "get": {
        "tags": [
          "AttendanceTrackerInfrastructure"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AttendanceTrackerInfrastructure/get-department/{departmentName}": {
      "get": {
        "tags": [
          "AttendanceTrackerInfrastructure"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "departmentName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AttendanceTrackerInfrastructure/add-admin": {
      "post": {
        "tags": [
          "AttendanceTrackerInfrastructure"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AdminAPI"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AttendanceTrackerInfrastructure/add-staff": {
      "post": {
        "tags": [
          "AttendanceTrackerInfrastructure"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StaffAPI"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AttendanceTrackerInfrastructure/add-workday-record": {
      "post": {
        "tags": [
          "AttendanceTrackerInfrastructure"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/WorkdayRecordAPI"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AttendanceTrackerInfrastructure/add-department": {
      "post": {
        "tags": [
          "AttendanceTrackerInfrastructure"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DepartmentAPI"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "AdminAPI": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DepartmentAPI": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "StaffAPI": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "department": {
          "type": "string"
        },
        "workdays": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkdayRecordAPI"
          }
        }
      },
      "additionalProperties": false
    },
    "WorkdayRecordAPI": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date",
          "type": "string"
        },
        "checkIn": {
          "format": "time",
          "type": "string"
        },
        "checkOut": {
          "format": "time",
          "type": "string"
        },
        "staffName": {
          "type": "string"
        },
        "staff": {
          "$ref": "#/definitions/StaffAPI"
        }
      },
      "additionalProperties": false
    }
  }
}